requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
requests

elevenlabs
openai>=1.0.0
azure-cognitiveservices-speech
google-cloud-texttospeech
pydub


boto3

ffmpeg-python

structlog
asgi-correlation-id
